backend/requirements.txt
# =============================================================================
# VimaBOT SaaS - Dependências Python
# =============================================================================
# Todas as bibliotecas necessárias para o backend da plataforma.
# Instalar: pip install -r requirements.txt
# =============================================================================

# --- Framework Web Assíncrono ---
fastapi==0.115.6
uvicorn[standard]==0.34.0

# --- ORM e Banco de Dados ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# --- Validação e Configuração ---
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0
python-dotenv==1.0.1

# --- Autenticação e Segurança ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# --- Cliente HTTP Assíncrono ---
httpx==0.28.1

# --- Armazenamento S3 (Minio) ---
aiobotocore==2.18.0

# --- Bot Engine Telegram ---
aiogram==3.17.0

# --- Utilitários ---
python-dateutil==2.9.0
